refactor(beta10): preserve runtime quality ratchets
