+++ docs/v4/implementation/beta10/rolling-compatibility-matrix.md
@@ -9,7 +9,7 @@ Compatibility is a decision per artifact family, not one `old/new` Boolean.
 | Wire               | Calls carry the exact application, client-contract and wire digests.                                              | An incompatible frame is rejected before the operation body; Mutations do not acquire compatibility by retry.                              | Round-robin POST construction and success assertion at `tests/load/beta10-ten-instance.ts:51`–`:68`.                                                                                                                                                            |
 | Policy and Context | A rolling instance uses the projections signed into its own Runtime Build.                                        | No fallback or cross-build Policy/Context interpretation is claimed.                                                                       | Runtime artifact exactness precedes readiness; this slice adds no compatibility bypass.                                                                                                                                                                         |
 | Realtime           | A reconnect to any instance of a compatible deployment re-authorizes and resumes or resets from PostgreSQL state. | Resume state is deployment-bound; retirement waits or the client receives a freshly authorized reset.                                      | Child-process crash and fresh-Runtime reconnect at `tests/integration/postgres/beta07-generated-live-query.test.ts:397`–`:447`; arbitrary-holder takeover re-authorizes at `tests/integration/postgres/beta07-postgres-no-affinity-carrier.test.ts:727`–`:787`. |
-| Executable         | A worker may claim only a run whose pinned executable digest it carries.                                          | An old run remains for a compatible worker; it is hidden from incompatible admission and a direct race still refuses `EXECUTABLE_RETIRED`. | Admission filters before `LIMIT` at `packages/runtime/src/durable/postgres-kernel.ts:473`–`:504`; claim checks again under the row lock.                                                                                                                        |
+| Executable         | A worker may claim only a run whose pinned executable digest it carries.                                          | An old run remains for a compatible worker; it is hidden from incompatible admission and a direct race still refuses `EXECUTABLE_RETIRED`. | Admission filters before `LIMIT` at `packages/runtime/src/durable/postgres-kernel.ts:357`–`:393`; claim checks again under the row lock.                                                                                                                        |
 | Durable state      | PostgreSQL transitions and leases are shared across all compatible instances.                                     | A stale holder loses its fence; a serialization loser performs no second transition.                                                       | Ten-instance zero-duplicate assertion at `tests/load/beta10-ten-instance.ts:79`–`:112` and one intentional lease recovery at `tests/load/beta10-soak-chaos.ts:55`–`:116`.                                                                                       |
 
 ## Exact claim
