+++ tests/goldens/beta01/generated-digests.json
@@ -14,13 +14,13 @@
 		"durable-kernel.json": "1710a5766e79c89f93a72b4819e755a41d2366c86c34d72f3b428f9d38359afa",
 		"execution-composition-explain.json": "0fd7f77cf49a1d372e5d41ff4acf104583072e9b1a11d27b589d9bbde011a6c8",
 		"field-normalizer-programs.json": "3c6a69af54d1a5fb273fa02db8114a3ddc80d2417fcb4a3f9ebc8f8bf8a78e82",
-		"internal/application-346fcckz.js": "35599ca13247534509ac1c34a46ffcf80de67b33e6c3fbe5bdde6778a8ed636f",
-		"internal/application-dta3f59z.js": "0fdc2d03260f4ae24c17e44811b3b26a47fe65e690905bfbc48afa6c54175a53",
-		"internal/application-t0wdqfrj.js": "61936c8a01b1e2dbd46f6524250552b14664a3766a9dbb099a4343bccd3aaa2e",
-		"internal/application-yt9f6kcv.js": "dbfc8a391b55bee7cf978b959811e8bc8bcd47c3ca3312c244f85c8fba29d873",
+		"internal/application-122br4k7.js": "e619388b73cb89eefa2033b4b03e6dfb66f7e54ade386d8d493418e13e52583b",
+		"internal/application-p4gjaxyj.js": "96e854378a3d432dbdfad25b23926349248b9cd4cfa0fc0a80268aa4be7ef710",
+		"internal/application-vqw92zgv.js": "7d0dc9b4d2681bd1d830d6901057b687b4867fe00ece574dadf930b78c9f96d3",
+		"internal/application-zbzmy0rb.js": "f899f601f862c2c1b2f60757a8e37aa87d7abad0311f2449f335e69fb7be9b40",
 		"internal/application.d.ts": "240b04a22aefd74f658fb6fb5da2bf44df98ea8edc3d95dfcbaa80aed5f400c8",
-		"internal/application.js": "271b567649972277f7ae5e81525b628c7bd7b9e086fee6b2dbf0cc64325b9b1d",
-		"internal/checksums.json": "429325e5b498f642ab66324fa56caa49bf49d578c40bccd34de6eb504257a155",
+		"internal/application.js": "67dee2eaa32008cc33eab44b7c4e5c2d76071006727516c19f441b5819cf2d35",
+		"internal/checksums.json": "b766adcecf88fe4a3e972ce11440290474cc49d2baa5ef431e49b04e707f096c",
 		"internal/package-contracts/questpie-collaboration-audit-846963f083917e90c9a1fa4c25d7ac12de3ef0dc1fb82b1c2badd14616c61c0c.ts": "ca79dff0c95d5c669e0a815288f69200977258d3ff5882d0b9f4e0e5f7ef110c",
 		"internal/package-inventories.json": "2ffe2e9595411fa40b2fbfa51b9d3c853bf4be9286a1135ac866f9428b54d12d",
 		"live-query-dependency-algebra.json": "cd0f12a8f6b6e1e89c6d028cb1a6d7352e320291d8eb849e13dac1b4ebbac297",
@@ -40,7 +40,7 @@
 		"realtime-wire-contract.json": "0eaa0a63b9dc52f59494c6f76664873eb8cbace875f9f80ab26895d28b4eacc2",
 		"relational-explain.json": "f379a4ad5a579d1d84bd51791a14da04c55281c26f5a5a6f3578cf6d915f9c89",
 		"relational-nondisclosure.json": "a50d3e0c1c4251540569d641044d65e55f164f153a67adf57aa2735c52f85833",
-		"runtime-build.json": "3826688d640c5f1eeffe75ec915f97a812c57b48a53e6eb7d045e75765481a6a",
+		"runtime-build.json": "0c874431c1582b2b0f41d76b540cc343cbf20b0e12a38fa440c9b3e76856c358",
 		"runtime-executables.json": "67e61ed5a5fe0f7ff6f040c494ed3b8f6bb9b1f5220654c877d6ddf2fb740fc0",
 		"schema-projection.json": "56b20c6134ee9f171ef6264202ef6fafe36da21bc17b77263939ddb48485f623",
 		"server-value-programs.json": "efa672374639def5c7a918fd43073168dac2974964238ef548cf2986c2b4dfa7",
