+++ tests/type/beta01-generated-contract.test.ts
@@ -86,10 +86,10 @@ describe("BETA-01 generated contract", () => {
 			"durable-kernel.json",
 			"execution-composition-explain.json",
 			"field-normalizer-programs.json",
-			"internal/application-346fcckz.js",
-			"internal/application-dta3f59z.js",
-			"internal/application-t0wdqfrj.js",
-			"internal/application-yt9f6kcv.js",
+			"internal/application-122br4k7.js",
+			"internal/application-p4gjaxyj.js",
+			"internal/application-vqw92zgv.js",
+			"internal/application-zbzmy0rb.js",
 			"internal/application.d.ts",
 			"internal/application.js",
 			"internal/checksums.json",
